Add files via upload
Hotel Booking Analysis.

In this data set contains booking information of two types of hotels that are city hotel and a resort hotel. This project will help us to understand the behavior of hotel industry like how it works. It includes the information such as when the booking was made, length of stay, the number of adults, children and babies, and the number of available parking spaces and many more. Hotel bookings depends on many factors and we are going to put light on some major factors which will help to improve the performance of hotel in terms of revenue,number of bookings etc.